Add initial unit tests
We don't spend time re-testing Joi validates that something is required and no greater than a certain length.

But we do have some logic in the translator that relies on ensuring a region is valid, and we will be making some changes. So, adding this test to provide a basline and ensure we don't break anything with our changes.