Rename bill run plugin to be more consistent
Since adding the `BillRunPlugin` we've added a plugin to add a notifier instance to the request (`RequestNotifier`) and another to add an invoice instance (`RequestInvoicePlugin`).

The services that these plugins depend on also have the same names, with the exception of the `BillRunPlugin` where the plugin differs from the service name (`RequestBillRunService`).

So, this is just a housekeeping change to rename `BillRunPlugin` to make it consistent with the other plugins.