Create rebilling initialise service (#395)
https://trello.com/c/GVybPen8/1945-m-receive-respond-to-re-billing-request

The next step is a service which will initialise two empty invoices and link them to the invoice to be rebilled. This PR implements this InvoiceRebillingInitialiseService. It accepts the bill run that the invoice is to be rebilled to along with the invoice being rebilled, and returns an object containing the cancel invoice (ie. the invoice which will cancel out the invoice being rebilled) and the rebill invoice (ie. the invoice which we will be using to rebill the customer), along with a response to be passed straight back to the caller.

While developing this we discovered an issue with the existing ON CONFLICT constraint, which we tracked down to the fact that Postgres does not see null in one entry as being equal to null in another; this led to multiple entries being created when the constraint should have stopped them. This PR therefore includes a db migration and model amendment to resolve this.