Create InvoiceRebillingCreateTransactionService (#402)
https://trello.com/c/qkWbpJlf/1962-m-create-re-billing-transactions

The next stage of rebilling is to create the InvoiceRebillingCreateTransactionService that can be passed a transaction and a licence, and it will create a duplicate transaction on the specified licence. It can optionally invert the transaction type, ie. a debit will become a credit and vice-versa. We also update the tallies at the bill run, invoice and licence level to ensure the figures are accurate.