Add auth scope to invoice rebill endpoint
https://trello.com/c/sMC4Ns4D

We have recently added the `PATCH /v2/{regimeId}/bill-runs/{billRunId}/invoices/{invoiceId}/rebill` endpoint as part of implementing a new feature that allows for an invoice to be rebilled (cancelled out by one new invoice and regenerated in another).

This work won't be complete before we go live with the rest of our other functionality but the endpoint would currently be available. Though we wouldn't expect our client systems to attempt to use it, just to be sure we are ensuring access to it is only possible by the admin user.

So, this change adds the admin auth scope to the route's options plus any tweaks needed to make this work.