Update bill run invoices controller tests
Had to make some changes to faciliate authenticating as an admin user rather than a system user. The key difference is moving the generation of the authToken and the stubbing of `JsonWebToken.verify()` to a beforeEach block for each section, rather than doing it once for the whole suite of tests.

We then also encountered an issue with the new admin user not being linked to the same regime as the bill run. This was because the helper was automatically creating a regime of its own. So, we also made some changes to the AuthorisedSystemHelper to faciliate being able to set the regime to use for a new admin authorised system.