Remove fields from copied rebilling transactions (#413)
https://trello.com/c/qkWbpJlf/1962-m-create-re-billing-transactions

During rebilling development we noticed that date fields in existing transactions (created_at, updated_at and created_by) were being duplicated in the new transactions, when these should in fact be generated fresh. This PR changes this.

For created_at and updated_by we simply set these as undefined and the db will populate the dates for us. For created_by, we need to provide the authorised system making the rebilling request, which therefore needs a change to be made to the controller and InvoiceRebillingService to pass this along.