Add database clean commands and script
First step in being able to reset the database is cleaning it of all data. Locally, in testing we have found the reason a rollback may fail is because existing data causes an issue, for example, due to a constraint that is no longer valid.

So, first step for the reset is to clean the database of all data.