Add reset commands
By adding the reset commands we now have the means of reset a database to an empty and ready state.

We are still relying on rollbacks because there are actions taken in the migrations that would not be covered by a simple drop of all the tables. We add things like functions, relationships and sequences as part of the migrations. So, it's cleaner to rollback than just drop stuff.