Add auth scope to invoice rebill endpoint (#411)
https://trello.com/c/sMC4Ns4D

We have recently added the `PATCH /v2/{regimeId}/bill-runs/{billRunId}/invoices/{invoiceId}/rebill` endpoint as part of implementing a new feature that allows for an invoice to be rebilled (cancelled out by one new invoice and regenerated in another).

This work won't be complete before we go live with the rest of our other functionality but the endpoint would currently be available. Though we wouldn't expect our client systems to attempt to use it, just to be sure we are ensuring access to it is only possible by the admin user.

So, this change adds the admin auth scope to the route's options plus any tweaks needed to make this work.

** Notes - bill run invoices controller tests

Had to make some changes to facilitate authenticating as an admin user rather than a system user. The key difference is moving the generation of the authToken and the stubbing of `JsonWebToken.verify()` to a beforeEach block for each section, rather than doing it once for the whole suite of tests.

We then also encountered an issue with the new admin user not being linked to the same regime as the bill run. This was because the helper was automatically creating a regime of its own. So, we also made some changes to the AuthorisedSystemHelper to facilitate being able to set the regime to use for a new admin authorised system.