Reuse DatabaseHelper to clean
We pinched the functionality to clean the DB from the test `DatabaseHelper`. This is because we thought the helpers would not be available to us because the `test/` folder is not included when building the production docker image.

Then we remembered, we do! Specifically, the `test/support` folder as this was needed to support our test generate bill run endpoint.

So, this change updates the existing script to make use of it and remove the duplication in our code.