Add db reset package.json script (#414)
We have the ability to roll back migrations through the existing scripts. We can then follow that up with a reseed of the DB.

Locally that is relatively simple to do through chaining the commands we pass through (see the VSCode task `🗄️ DB (CMA)` for an example).

But we could also do with this ability in our AWS environments (in non-production) to support testing. Ideally, to make setting up the Jenkins job easier this would be kicked off from a single `npm run` script.

So, this change implements that run script and updates the existing VSCode tasks to use it.

** Notes

Locally, in testing, we have found the reason a rollback may fail is that existing data causes an issue, for example, due to a constraint that is no longer valid. So as part of this change, we have added the ability to clean out all existing data before attempting a rollback.

We are still relying on rollbacks because there are actions taken in the migrations that would not be covered by a simple drop of all the tables. We add things like functions, relationships and sequences as part of the migrations. So, it's cleaner to rollback than just drop stuff.