Add check to prevent updating the admin
We need to ensure no changes can be made to the main admin authorised system we seed when the service first goes live.

We depend on this account for administrating the service so cannot afford to make changes to it that might lock us out. So, we block all updates to it as a safety measure.

Should an update be needed, we'll need to do it through a migration and an RfC.