Prevent duplicate billing (#418)
https://trello.com/c/e2Z8puRX/1976-prevent-duplicate-re-billing-of-a-single-invoice

In order to prevent additional rebilling, we need to add validation to check whether it is currently being rebilled. We do this by checking to see if any invoices exist which have the invoice's id in the rebilled_invoice_id field.

This also satisfies the criteria that an invoice that has been rebilled can be rebilled again if the previous rebilling has been cancelled; in this scenario, both the cancel invoice and the rebill invoice will have been deleted meaning there won't be any invoices with the id in the rebilled_invoice_id field.