Add ability to update regimes
This is the final part of the update process and allows for an authorised system's regimes to be updated. The idea is that the regimes will be set based on whatever is in the payload. So, if you wanted to add a regime you would need to also send the 'slug' for the existing regime the authorised system is linked to.