Add support to seed test authorised system (#420)
Currently, we support seeding the 'admin' and a 'system' authorised system (user).

What has actually been happening is we have been seeding the 'admin', but using WRLS's credentials as the 'system' user. We're getting close to going live with the service so we need to move away from using WRLS's credentials and use our own system account.

Whilst looking into what accounts we have we noted there is one earmarked for the TCM, though it is no longer our intent to link the 2 systems. This observation came at the same time our test team asked if there was an account they could make changes to as part of their automated testing.

So, this change is all about adding support to seed a 'test' authorised system. But it's also signalling our intent to use our own credentials rather than our clients going forward.