Add update (PATCH) authorised systems endpoint (#416)
In [V1 of the API](https://github.com/charging-module-api) there was an endpoint that allowed the admin user to change the details of an 'authorised system' (our name for user accounts!)

We've always had the intention of bringing it into V2, it was just very low on the list of priorities. Now it's been identified as a need by our test team to allow them to automate the testing of authentication and authorisation.

So, this change adds a new update authorised system endpoint which will allow us to

- amend the status of an authorised system
- amend the name of an authorised system
- amend what regimes an authorised system has access to

Note - Included in the validation are checks to ensure the main 'admin' account cannot be amended. This is to ensure we don't accidentally/maliciously lock ourselves out of the service.