Include pending changes in SendCustomerFileService (#425)
https://trello.com/c/MAhCxVK6

One of the issues we noted was that should a customer file fail to arrive in the S3 bucket we have no means to recover and re-try sending the it.

So, if customer changes had been added we _could_ process them and generate a `customer_file` record with a generated customer file reference. These are wrapped in a transaction so assuming that was successful the database will have been updated. But if, for example, an issue occurred with accessing AWS S3 the file would not get transferred to SSCL.

The changes would become 'stuck'. Processed and ready to go but no way to kick of the file generation again. After reviewing the situation with our BA, we agreed the other issue is that this could go unnoticed for some time. As it is can be the case that no changes are needed for a regime and bill run in a given week it would be normal for SSCL to see no file arrive. From the client's perspective, no errors or errors would have been returned when they sent the corresponding bill run.

Therefore we agreed that the existing send customer file process should try to cater for both new changes and any old ones that have yet to be sent.

This change adds a new `PrepareCustomerFileService` to focus on preparing new customer changes for sending. It then updates the existing `SendCustomerFileService` to check for and send _any_ `customer files` with a status of `pending`, which includes both new and old ones.

Note - The bulk of the work was done in a pairing session. Hence, we are comfortable this is a big change not broken down for review.