Add nuke-db npm script
When migrations go bad, they go _real_ bad! Sometimes, when working locally our only solution is to drop the whole environment and start again.

As an interim step, we'd like the option of being able to simply 'nuke' the DB. Essentially drop it completely and start again.

This change adds a `package.json` script that will allow us to do that.