Add drop database script
Using the existing `create.database.js` as a template, we update it to allow us to drop the database. The key point is like `create.database.js` the commands need to be run from the main `postgres` database. We cannot connect to `sroc_charge_db` and then attempt to drop it whilst connected!