Revised Admin Send Transaction File endpoint (#429)
https://trello.com/c/ntPvucIj/1974-enable-admin-endpoint-to-manually-trigger-a-transaction-file-v2

After initially raising a PR to create the endpoint and a stub AdminSendTransactionFileService, we realised once we started work on developing the admin service that this wouldn't be necessary -- we can simply call the existing SendTransactionFileService from the admin controller. We therefore pulled the review request for the previous PR and raised this one to create the admin endpoint and call SendTransactionFileService from within it.

While manually checking the endpoint, we found that the request would be rejected by the RequestBillRunService plugin before reaching the controller; this was due to the bill run status being pending which is rejected by the plugin's validation. We have therefore changed the plugin's validation so the 'editable' check is not performed if an /admin/ path is requested.