Rename migrations to control ordering
Initial tests threw up ann issue with the existing order due to dependencies thhat were covered by the alterations but not when just creating a table for the first time. For example, transactions referencing licences which as yet don't exist.

So, taking a timestamp from today this renames the migrations to allow us to control the order they will be run.