Squash migration scripts before Go Live (#432)
The period of greatest change in the DB migration scripts is typically in the early days of a project as you are iterating the core functionality and features. Most scripts are added during this time. At the time of writing, we have **47** of them!

Some teams view them as sacrosanct and never touch them (some just don't have the time!) We, however, like to take the opportunity before a service goes live to 'squash' them down. This means merging all the alterations made to tables back into their original 'create' scripts. The idea is you end up with only `create_` and `add_` migration scripts before 'go live'.

The benefit to us as a delivery team is

- fewer files!
- less code!
- less chance of complications with rollbacks due to the number and interdependance of changes

Once the service goes live, we'll also switch to the principle that migrations should never be touched. But right now, we're using this window to get a better handle on them.

** Notes - Initial tests threw up an issue with the existing order due to dependencies that were covered by the alterations but not when just creating a table for the first time. For example, transactions referencing licences which as yet don't exist.

So, taking a timestamp from today this change also renames the migrations to allow us to control the order in which they will be run.