Add handling of 'root' routes
We started updating the route files directly and spotted that the root endpoints are a special case. We specifically add an options property where we set `auth` to `false`. This tells Hapi not to authenticate any requests to these endpoints.

Our current logic didn't cater for that hence we need to add the test back in but tweaked to expect the alternate `options` property.