Avoid modifiying the passed in property
We fell foul of this in the tests. When we just passed in `baseRoute` into the service for the root unit tests other tests would fail because they had modified the value by reference. So, we catered for this by setting up a test `rootRoute`.

But we also want to try and avoid modifing things by reference if we can.