Remove references to auth options in route files
We expect _all_ the tests to break with this one! But now all this will be handled by the service and the router plugin which we'll soon update to use it.

Another thing to note is we do lose descriptions. But as we had not been consistent with setting them, and we have our swagger API its something we feel we can live with.