Update RouterPlugin to use service
Doing this should have got all the tests passing again. If you look at the startup output from 'blipp' everything is as expected.

However, the admin `BillRunsController` test fails. A quick check has found we've made a mistake in a previous commit by _not_ adding the auth scope to this endpoint and then writing a test that uses a system user.

The new service has correctly determined this _should_ have the admin auth scope which is why the tests are now failing. We're gonna fix the issue in another PR and once that is approved and merged, we can rebase this and all should be working.