Be consistent with our approach
We had an alternative to this filtering approach where everything was done in one place, specifically we controlled in one service which routes got what auth scope. The idea was we could protect unfinished endpoints by hiding them behind the admin scope.

It worked, but it meant stripping out all the options config we had added to the route files. This didn't sit well with us because it meant you could no longer look at a route and understand its full config, and whether it was an admin route, a standard one, or didn't require authentication at all.

But if this was our guiding principle, we were rightly challenged on putting the config to determine if a route gets added or not in one place. Again, wouldn't it be better to also have that in the routes files?

It was a brilliant observation and very correct. So, using that feedback this change updates the solution to look out for that custom app config item on a route to determine whether to filter it.