Fix generated bill run status after add trans.
We've spotted an edge case the CM isn't handling. Once a bill run is `generated` we allow 2 kinds of edits

- deleting an invoice
- adding a transaction

Our commitment to client systems is that the invoice and credit note values of a generated bill run are correct. When deleting an invoice, we are able to (and appear to be!) correctly recalculating the invoice and credit note values to reflect the change.

Adding a transaction is harder. What was an invoice could suddenly become a credit note, a zero value transaction, or something even more complicated like minimum charge. Because of this the expectation was that the status would have to go back to `initialised` and the client system would be required to call the `/generate` endpoint again.

A recent test has highlighted we're not doing that. We are leaving the status as `generated` which means the values are incorrect. We are breaking that commitment to our client systems.

This change fixes the issue, by ensuring we set the bill run status to initialised (and reset the invoice and credit note values) whenever a transaction is added.