Fix Client ID unique constraint for rebilling
https://trello.com/c/qkWbpJlf

We have a constraint on the `transaction` record to ensure we don't allow 2 transactions with the same `regime_id` and `client_id` to be inserted. This is how client systems have confidence we don't allow duplicate transactions to get into an invoice.

We forgot about this though when adding the rebilling functionality. All our manual and unit testing did not include having an original invoice which featured transactions with client ID's set.

In recent testing if we have found that the copying of transactions from the original to the rebilling invoices fails because of this constraint.

So, this change fixes the issue to ensure we still prevent duplicate transactions being added but allow us to create rebilling invoices.