Remove 'approved' as an editable bill run type
We have faced questions recently about what we should do when amending a bill run that has status of `approved`. Within the team and after speaking to WRLS we are all agreed that once a bill run is marked as `approved` it should not be possible to edit it. If changes are needed the client system _must_ delete the bill run and start again.

This change updates the editable stats to remove `approved.