Bring back $editable()
We started re-writing the tests in `app/services/plugins/request_bill_run.service.js` with the expectation that what we'd need to add alongide `$patchable()` is `$deleteable()`. But this overlooked adding a transaction which comes as a result of a `post` request not `delete`.

The permitted bill run status for deleting an invoice or adding a transaction is the same. So, this change brings back `$editable()` only no with `approved` dropped as an editable status.