Fix Client ID unique constraint for rebilling (#448)
* Fix Client ID unique constraint for rebilling

https://trello.com/c/qkWbpJlf

We have a constraint on the `transaction` record to ensure we don't allow 2 transactions with the same `regime_id` and `client_id` to be inserted. This is how client systems have confidence we don't allow duplicate transactions to get into an invoice.

We forgot about this though when adding the rebilling functionality. All our manual and unit testing did not include having an original invoice which featured transactions with client ID's set.

In recent testing if we have found that the copying of transactions from the original to the rebilling invoices fails because of this constraint.

So, this change fixes the issue to ensure we still prevent duplicate transactions being added but allow us to create rebilling invoices.

* Add test to replicate scenario

* Implement solution

Needed to tweak the test. Looks like if a property is `null` in the db when you request it on the Objection instance you get `undefined` not `null` returned.

But other than that this appear to resolve the issue.

* Correct and update other tests

Overlooked that we have other tests that incorporate fields that are not duplicated when copying the rebilling transaction.

This updates those as needed.