Refactor bill run editable and add patchable (#449)
* Remove 'approved' as an editable bill run type

We have faced questions recently about what we should do when amending a bill run that has status of `approved`. Within the team and after speaking to WRLS we are all agreed that once a bill run is marked as `approved` it should not be possible to edit it. If changes are needed the client system _must_ delete the bill run and start again.

This change updates the editable stats to remove `approved.

* Rename editable() to patchable()

We also include a guard clause so we quit the validation straight away of the request is a `GET`. This should make our next changes a little clearer/easier.

* Bring back $editable()

We started re-writing the tests in `app/services/plugins/request_bill_run.service.js` with the expectation that what we'd need to add alongide `$patchable()` is `$deleteable()`. But this overlooked adding a transaction which comes as a result of a `post` request not `delete`.

The permitted bill run status for deleting an invoice or adding a transaction is the same. So, this change brings back `$editable()` only no with `approved` dropped as an editable status.

* Refactor of RequestBillRunService

This covers re-writing the tests to try and cover each scenario where you can and cannot make changes to a bill run.

* Fix broken test

We've amended the error messages to distinguish between when you attempt to patch and unpatchable bill run and when you attempt to edit and uneditable bill run.

* Attempt to fix sonarcloud issue

> Refactor this function to use "return" consistently.

* Update some code comments

Co-authored-by: Stuart Adair <43574728+StuAA78@users.noreply.github.com>