Create ValidateBillRunLicenceService (#445)
https://trello.com/c/ZjWhOdbQ/1984-remove-licences-initial-validation

This service is passed a licence, a bill run id and an invoice id, and validates that the licence belongs to the bill run and that the invoice it belongs to is not a rebilling invoice (ie. it is an original invoice, so its rebilledType is O).

As part of this we add an $originalInvoice method to InvoiceModel which we use to check if it's a rebilling invoice or not.