Link rebilled transaction to original transaction (#456)
https://trello.com/c/qkWbpJlf

As part of the rebilling functionality, we record a link between the new rebilled invoices (`C` and `R`) and the original (`O`). We do this by recording the original invoice ID in `invoices.rebilled_invoice_id`.

We have also been asked to record a link between the rebilled transactions and the original transaction. This change adds that ability along with adding it to the `GET /invoice` response so it can be seen by client systems.