Housekeeping - Update lib file names
https://github.com/DEFRA/sroc-service-team/issues/68

A simple chore task to correct an oversight/niggle with the files in the `app/lib/` folder. All our others include the type in the name, for example `app/services/delete_file.service.js`. To avoid any confusion we want to also follow the same convention in the one place that currently doesn't!