Refactor InvoiceRebillingService to do more
https://trello.com/c/A91YBKNC

As part of adding the fix to [protect bill run when rebilling](https://github.com/DEFRA/sroc-charging-module-api/pull/461) we've realised we need to do more work in the controller.

Specifically, we need to manage the status of the bill run, setting it to `pending` when the process starts and resetting it afterwards. We're already doing more than we would typically do in a controller handler for rebilling. So, it's time to do a spot of refactoring.

The first step involves updating `InvoiceRebillingService` to manage more of the process by moving the `InvoiceRebillingInitialiseService` into it. As part of that we want to move what it currently does into its own service. `InvoiceRebillingService` can then become responsible for managing the process and in a future change, dealing with the status of the bill run.