Remove presenter handling from initialise service
We only really had this here because we need to prep the response before starting the copy process and we didn't want to introduce a service just for that.

Now the InvoiceRebillingInitialiseService can focus on just initialising the 2 new invoices and leave dealing with the response to our rebilling service.