Recalculate invoice on delete licence (#453)
https://trello.com/c/zYnTXa36/1986-remove-licences-recalculate-invoice-summary-info

When we delete a licence, we want to recalculate the summary values of the invoice it belongs to:
* invoice.deminimis_invoice is set depending on whether or not deminimis now applies;
* invoice.zero_value_invoice is set depending on whether or not zero value now applies;
* the invoice is deleted if it now contains no licences;
* invoice.minimum_charge_invoice is set depending on whether or not there are any minimum charge transactions under the invoice;
* the summary counts and values are adjusted accordingly (note that this is not currently part of the ACs but we do this for deleting invoices so it stands to reason we also want to do this for deleting licences).

We intend to use the existing DeleteInvoiceService when we delete an empty invoice, since this will automatically handle the bill run-level counts for us. However as this is not strictly part of this task's ACs, this will be done when we come to look at updating bill run-level stats; for now, we have a TODO comment in the code to remind us of this.