Set bill run status to pending when initialising
Having double checked what is happening when we start creating the new transactions we have confirmed that the bill run model `patchTally()` method is resetting the bill run, which means putting the status back into an `initialised` state.

So, we thought we would have to retain the previous status but it turns out this is not the case. We know it's always going to get reset so we can just have the `InvoiceRebillingInitialiseService` set the bill run status to `pending` as that seems a good fit for actions taken under the banner of initialising.