Fix - protect bill run when rebilling (#464)
The WRLS team have spotted an issue with the rebilling process. When the `PATCH /rebill` endpoint is hit we generate the invoices and return their ID's. The process of copying the transactions and populating the new invoices happens in the background. As far as the client system is concerned the response is complete.

The issue that's been spotted is what happens if the client system was immediately to hit the `PATCH /generate` endpoint after rebilling an invoice? Currently, disaster! As long as at least one transaction has been created the generate process will kick off and create a summary for the bill run based on what is there at the time.

So, we need a solution. This change will update the rebill process to set the bill run's status as `pending`. Whilst the status is `pending` no other actions can take place against the bill run. We've used this convention for the `/send` transaction file and customer file processes. We also intend to apply it for any new changes. In retrospect (and we'll go back and update it at some time) `pending` should be the status we use when generating a bill run.

This then ties up with our existing bill run `GET /status` endpoint; it will always tell you if the data in a bill run is complete or if events are taking place that might mean you have to wait a sec.

** Notes - Having double-checked what is happening when we start creating the new transactions we have confirmed that the bill run model `patchTally()` method is resetting the bill run, which means putting the status back into an `initialised` state.

So, we thought we would have to retain the previous status but it turns out this is not the case. We know it's always going to get reset so we can just have the `InvoiceRebillingInitialiseService` set the bill run status to `pending` as that seems a good fit for actions taken under the banner of initialising. When the transactions have completed being added the DB transaction should complete and the bill run status be set to `initialised`.