Recalculate bill run on delete licence (#465)
https://trello.com/c/ntirSF2E/1987-remove-licences-recalculate-bill-run-summary-info

The final thing we want our delete licence service to do is to update the bill run values when a licence is deleted, which we do in this change.

We also cover the scenario where the invoice has no licences after the requested licence is deleted; in this case, we delete the invoice using DeleteInvoiceService to ensure that the bill run-level stats are updated for this.

Note that this required a small refactor to DeleteInvoiceService so that it when calling it in this way, we can pass in the current db transaction. However if it is being called from elsewhere (and therefore no transaction is passed in) then we instead create a new one.